usage: Document clone arguments (optional directory)
'dat clone' accepts an optional directory argument since 4fad9f4 (dat
clone works more like git clone now, 2014-04-10).  Mention that in the
usage docs.

I'm not very familiar with the range of JavaScript libraries, but if
there is something similar to Python's argparse module that lets you
declare your interface and then automatically constructs the parser
and usage docs, that would make it easier to keep the parser and docs
in sync.