add references to change index to data keys
By adding a reference to the change index id to data table keys, we can quickly
revert the current table of the repository to a particular checkpoint.

These changes only enable the storage of the change index keys.

This is not a stable commit. A majority of tests now pass, but there are still
significant issues.