store change ids alongside data to enable fast rollback
The addition of change log ids to the data keys (after the versions), will
allow us to quickly extract the state of the data at particular point in the
change log. This can be accomplished via a linear scan of the keys in the data
table, requiring that the change id of a particular object is <= the target
point in the log.

If we did not include this data alongside the data, we would be forced to
complete a reconstruction of the dataset via the change log. This would
complicate the process of rolling back particular subsets of the data to
predetermined points in the history. Additionally, it wouldn't be possible to
quickly determine the relative age of two objects, which has a number of
possible applications in reproducibility and logging.

The level-dat backend will support these change ids as of 4.5.0.

No functionality is yet tested which is based on the change ids, but the next
step should be to implement a commit/checkout or checkpoint/rollback model on
top of it.

With this update we now pass 616/616 tests.